backend/requirements.txt
alembic==1.17.2
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.6.0
anyio==4.11.0
appnope==0.1.4
asttokens==2.4.1
attrs==25.4.0
authlib==1.3.1
bcrypt==4.3.0
blinker==1.9.0
cachetools==5.3.3
certifi==2024.2.2
cffi==2.0.0
charset-normalizer==3.3.2
click==8.3.1
comm==0.2.2
cryptography==46.0.3
debugpy==1.8.1
decorator==5.1.1
distlib==0.3.8
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
executing==2.0.1
fastapi==0.124.0
filelock==3.15.4
gemini-api==0.1.6
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.2
google-api-core==2.19.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-genai==1.52.0
google-generativeai==0.5.2
googleapis-common-protos==1.63.0
grpcio==1.63.0
grpcio-status==1.62.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.7
ipykernel==6.29.4
ipython==8.24.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.1
jupyter_core==5.7.2
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib-inline==0.1.7
narwhals==2.12.0
nest-asyncio==1.6.0
numpy==2.3.5
packaging==24.0
pandas==2.3.3
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.2.1
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==21.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.8.0
PyMuPDF==1.26.6
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.21
pytz==2025.2
pyzmq==26.0.3
referencing==0.37.0
requests==2.31.0
resend>=2.0.0
rpds-py==0.29.0
rsa==4.9
setuptools==69.5.1
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.44
stack-data==0.6.3
starlette==0.50.0
streamlit==1.51.0
stripe>=14.1.0
tenacity==9.1.2
toml==0.10.2
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.38.0
virtualenv==20.26.3
wcwidth==0.2.13
websockets==15.0.1
wheel==0.43.0
